refactor(core): break both module dependency cycles
The analyser reported two import cycles. Both existed because a large
module played two roles at once: defining a value everyone needs, and
aggregating types from its own dependents.

`src/core/search/types.ts` both declared `SearchError` and re-exported
leaf types from `property-filter`, `temporal-intent` and
`retrieval-receipts`, so the leaves imported the error back out of the
hub and closed the loop. The error vocabulary and the ranked-row shape
now live in leaf modules that import nothing; `types.ts` keeps its path
and re-exports them, so its 39-symbol export surface is unchanged and no
importer moves.

`src/core/brain/paths.ts` and `vault-identity.ts` were mutually
dependent over a single root-path constant. The vault-relative layout
constants now sit in their own leaf module, re-exported from `paths.ts`
with its 111-symbol surface intact.

Both surfaces were compared symbol-by-symbol against HEAD through the
TypeScript checker rather than by eye. A ratchet test walks the static
import graph and fails on any cycle, validated against an archived copy
of the pre-fix tree so it cannot pass vacuously.
