+++ src/core/search/time-range.ts
@@ -15,7 +15,7 @@
  * throws `SearchError("INVALID_INPUT")` — explicit, never silent.
  */
 
-import { SearchError } from "./types.ts";
+import { SearchError } from "./search-error.ts";
 
 const HOUR_MS = 60 * 60 * 1000;
 const DAY_MS = 24 * HOUR_MS;
